chore: test react-resval in prod
