+++ docs/docs-react/src/app/app.tsx
@@ -2,6 +2,7 @@ import { Route, Routes, Link } from 'react-router-dom'
 import Home from './home'
 import About from './about'
 import TestField from './__test__'
+import V1 from './v1'
 
 export default function App() {
   return (
@@ -28,8 +29,17 @@ export default function App() {
         path='/__test__'
         element={
           <div>
-            <Link to='/'>Test Field</Link>
+            <Link to='/'>Home</Link>
             <TestField />
+          </div>
+        }
+      />
+      <Route
+        path='/v1'
+        element={
+          <div>
+            <Link to='/'>Home</Link>
+            <V1 />
           </div>
         }
       />
