+++ packages/react-resval/src/v1/react-resval.ts
@@ -78,34 +78,34 @@ export default createResponsiveValues
  *
  */
 
-const useVx = createResponsiveValues({
-  breakpoints: {
-    base: '0px',
-    xs: '320px',
-    sm: '576px',
-    md: '768px',
-    lg: '1080px',
-    xl: '1280px',
-  },
-  media: 'min',
-})
+// const useVx = createResponsiveValues({
+//   breakpoints: {
+//     base: '0px',
+//     xs: '320px',
+//     sm: '576px',
+//     md: '768px',
+//     lg: '1080px',
+//     xl: '1280px',
+//   },
+//   media: 'min',
+// })
 
-const utilityValues = {
-  fontSize: {
-    base: '12px',
-    xl: '24px',
-  } as const,
-  color: {
-    base: 'red',
-    '600px': 'blue',
-    xl: 'green',
-  } as const,
-}
+// const utilityValues = {
+//   fontSize: {
+//     base: '12px',
+//     xl: '24px',
+//   } as const,
+//   color: {
+//     base: 'red',
+//     '600px': 'blue',
+//     xl: 'green',
+//   } as const,
+// }
 
-// eslint-disable-next-line react-hooks/rules-of-hooks
-const { fontSize, color } = useVx(utilityValues)
+// // eslint-disable-next-line react-hooks/rules-of-hooks
+// const { fontSize, color } = useVx(utilityValues)
 
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-const fontSizeProp: '12px' | '24px' | undefined | null = fontSize
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-const colorProp: 'red' | 'green' | 'blue' = color
+// // eslint-disable-next-line @typescript-eslint/no-unused-vars
+// const fontSizeProp: '12px' | '24px' | undefined | null = fontSize
+// // eslint-disable-next-line @typescript-eslint/no-unused-vars
+// const colorProp: 'red' | 'green' | 'blue' = color
